requirements.txt
discord==2.2.3
openai==0.27.0
retry==0.9.2
python-dotenv==0.19.1